js/xpconnect/tests/chrome/test_doublewrappedcompartments.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=533596
-->
<window title="Mozilla Bug 533596"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">

  <iframe type="content"
          src="http://example.org/tests/js/xpconnect/tests/mochitest/file_doublewrappedcompartments.html"
          onload="go()"
          id="ifr">
  </iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
      const Ci = Components.interfaces;
      const utils = window.QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIDOMWindowUtils);

      function go() {
        var wrappedWin = $('ifr').contentWindow;
        var unwrapped = wrappedWin.wrappedJSObject;

        var filter = unwrapped.filter;
        is(utils.getClassName(filter), 'Proxy', 'properly wrapped');
        is(typeof filter.QueryInterface, 'function', 'double wrapped');

        ok(unwrapped.testme(filter),
           "content didn't get a proxy, but another double wrapped object");
        SimpleTest.finish();
      }

      SimpleTest.waitForExplicitFinish();
  ]]></script>
</window>
